backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask==3.0.3",
    "pydantic>=2.12.5",
    "pytest==8.3.2",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.7.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ty>=0.0.15",
]
